Enemies Drawn!
Enemies are now drawn on screen and have corners set. Next I need to
work on bullets firing and colliding with enemies. After that get the
score counter to go up with each kill and then do the menu navigation
system. Then this assessment is done and ready to go!